Fixed a few bug related to memory based LRUCache

+++ lib/classes/cache/inMemory/LRUCache.h
@@ -18,13 +18,13 @@
 @interface LRUCache : BaseCache
 
 /*!
-The maximum number of elements that can be stored in memory.
+ The maximum number of elements that can be stored in memory.
  */
 @property (nonatomic, assign) NSInteger maxElementsInMemory;
 
 /*!
-The maximum memory that can be used to store elements
+ The maximum memory that can be used to store elements
  */
-@property (nonatomic, assign) NSInteger maxMemoryAllocated;
+@property (nonatomic, assign) float maxMemoryAllocated;
 
 @end
